+++ cmd/echo_test.go
@@ -3,7 +3,11 @@ package cmd
 import (
 	"fmt"
 	"os"
+	"tcping2/test"
 	"testing"
+
+	"github.com/ory/dockertest/v3"
+	log "github.com/sirupsen/logrus"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/tommi2day/gomodules/common"
@@ -12,11 +16,28 @@ import (
 var echoPort = common.GetEnv("ECHO_PORT", "39207")
 var echoHost = common.GetEnv("ECHO_HOST", "127.0.0.1")
 var echoIP = common.GetEnv("ECHO_SERVER", "0.0.0.0")
+var echoContainer *dockertest.Resource
+
+func TestMain(m *testing.M) {
+	var err error
+
+	test.InitTestDirs()
+	if os.Getenv("SKIP_ECHO_SERVER") != "" {
+		return
+	}
+	echoContainer, err = prepareEchoContainer()
+	if err != nil {
+		log.Fatalf("prepareEchoContainer failed: %s", err)
+	}
+	code := m.Run()
+	destroyEchoContainer(echoContainer)
+	os.Exit(code)
+}
 
 func TestEchoClient(t *testing.T) {
 	var err error
 	var out string
-
+	test.InitTestDirs()
 	t.Run("Standard Server", func(t *testing.T) {
 		unitTestFlag = true
 		args := []string{
@@ -37,6 +58,7 @@ func TestEchoClient(t *testing.T) {
 		if os.Getenv("SKIP_ECHO_SERVER") != "" {
 			t.Skip("Skipping Echo Server tests")
 		}
+
 		queryPort = echoPort
 		unitTestFlag = true
 		args := []string{